# docs/swagger.yaml
basePath: /
definitions:
...
  managers.Response:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  managers.ResponseID:
    properties:
      code:
        type: integer
      id:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
...
  six910-database-interface.Product:
    properties:
      color:
        type: string
      cost:
        type: number
      currency:
        type: string
      depth:
        type: number
      desc:
        type: string
      distributorId:
        type: integer
      dropship:
        type: boolean
      entered:
        type: string
      freeShipping:
        type: boolean
      gtin:
        type: string
      height:
        type: number
      id:
        type: integer
      image1:
        type: string
      image2:
        type: string
      image3:
        type: string
      image4:
        type: string
      manufacturer:
        type: string
      map:
        type: number
      msrp:
        type: number
      multibox:
        type: boolean
      name:
        type: string
      parentProductId:
        type: integer
      price:
        type: number
      promoted:
        type: boolean
      salePrice:
        type: number
      searchable:
        type: boolean
      shipSeperately:
        type: boolean
      shippingMarkup:
        type: number
      shortDesc:
        type: string
      size:
        type: string
      sku:
        type: string
      specialProcessing:
        type: boolean
      specialProcessingType:
        type: string
      stock:
        type: integer
      stockAlert:
        type: integer
      storeId:
        type: integer
      thumbnail:
        type: string
      updated:
        type: string
      visible:
        type: boolean
      weight:
        type: number
      width:
        type: number
    type: object
...
host: localhost:3002
info:
  contact:
    name: API Support
    url: http://www.ulboralabs.com/contact/form
  description: This is the Six910 (six nine ten) API
  license:
    name: GPL-3.0
    url: https://github.com/Ulbora/Six910/blob/master/LICENSE
  termsOfService: https://github.com/Ulbora/Six910/blob/master/tos.html
  title: Six910 API
  version: "1.0"
paths:
...
  /rs/product/add:
    post:
      consumes:
      - application/json
      description: Adds a new product to a store
      parameters:
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/six910-database-interface.Product'
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/managers.ResponseID'
      summary: Add a new product
      tags:
      - Product
  /rs/product/delete/{id}/{storeId}:
    delete:
      consumes:
      - application/json
      description: Delete a products from the store
      parameters:
      - description: products id
        in: path
        name: id
        required: true
        type: string
      - description: store storeId
        in: path
        name: storeId
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/managers.Response'
      summary: Delete a products
      tags:
      - Product
  /rs/product/get/category/{catId}/{storeId}/{start}/{end}:
    get:
      consumes:
      - application/json
      description: Get list of products for a store
      parameters:
      - description: category id
        in: path
        name: catId
        required: true
        type: string
      - description: store storeId
        in: path
        name: storeId
        required: true
        type: string
      - description: start index 0 based
        in: path
        name: start
        required: true
        type: string
      - description: end index
        in: path
        name: end
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/six910-database-interface.Product'
            type: array
      summary: Get list of products by category
      tags:
      - Product
  /rs/product/get/id/{id}/{storeId}:
    get:
      consumes:
      - application/json
      description: Get details of a product
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: string
      - description: store storeId
        in: path
        name: storeId
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/six910-database-interface.Product'
      summary: Get details of a product by id
      tags:
      - Product
  /rs/product/get/list/{storeId}/{start}/{end}:
    get:
      consumes:
      - application/json
      description: Get list of products for a store
      parameters:
      - description: store storeId
        in: path
        name: storeId
        required: true
        type: string
      - description: start index 0 based
        in: path
        name: start
        required: true
        type: string
      - description: end index
        in: path
        name: end
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/six910-database-interface.Product'
            type: array
      summary: Get list of products
      tags:
      - Product
  /rs/product/get/name/{name}/{storeId}/{start}/{end}:
    get:
      consumes:
      - application/json
      description: Get list of products for a store
      parameters:
      - description: product name
        in: path
        name: name
        required: true
        type: string
      - description: store storeId
        in: path
        name: storeId
        required: true
        type: string
      - description: start index 0 based
        in: path
        name: start
        required: true
        type: string
      - description: end index
        in: path
        name: end
        required: true
        type: string
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/six910-database-interface.Product'
            type: array
      summary: Get list of products by product name
      tags:
      - Product
  /rs/product/update:
    put:
      consumes:
      - application/json
      description: Update product data
      parameters:
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/six910-database-interface.Product'
      - description: apiKey required for non OAuth2 stores only
        in: header
        name: apiKey
        type: string
      - description: store name
        in: header
        name: storeName
        required: true
        type: string
      - description: store localDomain
        in: header
        name: localDomain
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: OAuth2 client ID only for OAuth2 stores
        in: header
        name: clientId
        type: string
      - description: User ID only for OAuth2 stores
        in: header
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/managers.Response'
      summary: Update a product
      tags:
      - Product
...
swagger: "2.0"